requirements.txt
Pillow~=10.0.0
python-ffmpeg~=2.0.10
pytest~=7.0.1